Extract all remaining game logic into the game management service
